Move core logic to sidkick worker
